Prediction Model/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.11
Flask==3.1.0
fonttools==4.56.0
graphviz==0.20.3
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
Werkzeug==3.1.3
xgboost==3.0.0
